Pequenas atualizações no TaskModel, TaskRepository e no UserService

+++ src/main/java/com/jadson/todosimple/repository/TaskRepository.java
@@ -3,16 +3,14 @@ package com.jadson.todosimple.repository;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import com.jadson.todosimple.models.Task;
 
 @Repository
-public interface TaskRepository  extends JpaRepository<Task, Long>{
+public interface TaskRepository  extends JpaRepository<Task, Integer>{
     
-        List<Task> findByUser_Id (Long id);
+        List<Task> findByUser_Id (Integer id);
 
     // @Query(value = "SELECT t FROM Task WHERE t.User.id = :id")
     // List<Task> findByUser_Id (@Param("id")Long id);
